+++ src/Meta_Index.cpp
@@ -773,7 +773,7 @@ struct Meta_Index::impl : public Meta_Index_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -1340,7 +1340,8 @@ bool Meta_Index::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Index::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Index::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -1452,7 +1453,8 @@ void Meta_Index::impl::validate( unsigned state, bool is_internal, validation_er
    // [<finish validate>]
 }
 
-void Meta_Index::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Index::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1957,7 +1959,7 @@ void Meta_Index::clear( )
    p_impl->clear( );
 }
 
-void Meta_Index::validate( unsigned state, bool is_internal )
+void Meta_Index::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
