+++ src/Meta_Initial_Record.cpp
@@ -449,7 +449,7 @@ struct Meta_Initial_Record::impl : public Meta_Initial_Record_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -826,7 +826,8 @@ bool Meta_Initial_Record::impl::value_will_be_provided( const string& field_name
    return false;
 }
 
-void Meta_Initial_Record::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Initial_Record::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -868,7 +869,8 @@ void Meta_Initial_Record::impl::validate( unsigned state, bool is_internal, vali
    // [<finish validate>]
 }
 
-void Meta_Initial_Record::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Initial_Record::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1255,7 +1257,7 @@ void Meta_Initial_Record::clear( )
    p_impl->clear( );
 }
 
-void Meta_Initial_Record::validate( unsigned state, bool is_internal )
+void Meta_Initial_Record::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
