+++ src/Meta_Permission.h
@@ -168,7 +168,7 @@ class META_PERMISSION_DECL_SPEC Meta_Permission : public class_base
 
    void clear( );
 
-   void validate( unsigned state, bool is_internal );
+   void validate( uint64_t state, bool is_internal );
    void validate_set_fields( std::set< std::string >& fields_set );
 
    void after_fetch( );
