+++ src/Meta_View.cpp
@@ -1137,7 +1137,7 @@ struct Meta_View::impl : public Meta_View_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -2270,7 +2270,8 @@ bool Meta_View::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_View::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_View::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -2354,7 +2355,8 @@ void Meta_View::impl::validate( unsigned state, bool is_internal, validation_err
    // [<finish validate>]
 }
 
-void Meta_View::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_View::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -3020,7 +3022,7 @@ void Meta_View::clear( )
    p_impl->clear( );
 }
 
-void Meta_View::validate( unsigned state, bool is_internal )
+void Meta_View::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
