+++ src/Meta_View_Field.cpp
@@ -2258,7 +2258,7 @@ struct Meta_View_Field::impl : public Meta_View_Field_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -3540,7 +3540,8 @@ bool Meta_View_Field::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_View_Field::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_View_Field::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -3715,7 +3716,8 @@ void Meta_View_Field::impl::validate( unsigned state, bool is_internal, validati
    // [<finish validate>]
 }
 
-void Meta_View_Field::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_View_Field::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -4661,7 +4663,7 @@ void Meta_View_Field::clear( )
    p_impl->clear( );
 }
 
-void Meta_View_Field::validate( unsigned state, bool is_internal )
+void Meta_View_Field::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
